Move template context functions to templating module
